# payment-service-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Scan Results: payment-service",
    "version": "1.0",
    "description": "APIs discovered from payment-service (./test_samples)",
    "x-generated-by": "Universal Polyglot API Scanner v3.1",
    "x-generated-at": "2026-01-31T23:32:47.898244",
    "x-service-name": "payment-service"
  },
  "paths": {
...
    "/admin/users/{id}": {
      "delete": {
        "summary": "Custom-Router endpoint",
        "description": "Discovered from [complex_python_app.py]:91\n\nFramework: Custom-Router\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "delete_admin_users_id",
        "tags": [
          "Custom-Router",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\complex_python_app.py",
        "x-source-line": 91,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "financial: \\bbilling\\b",
          "admin: \\badmin\\b",
          "admin: \\binternal\\b",
          "Mutation: DELETE",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ]
      }
    },
    "/admin/users/{user_id}": {
      "delete": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:66\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "delete_admin_users_user_id",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 66,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "admin: \\badmin\\b",
          "Mutation: DELETE",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: user_id"
          }
        ]
      }
    },
...
    "/api/auth/login": {
      "post": {
        "summary": "Express endpoint",
        "description": "Discovered from [express_app.js]:23\n\nFramework: Express\nLanguage: JavaScript\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_auth_login",
        "tags": [
          "Express",
          "JavaScript"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\express_app.js",
        "x-source-line": 23,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\btoken\\b",
          "auth: \\bauth\\b",
          "auth: \\bjwt\\b",
          "pii: \\bemail\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [flask_app.py]:28\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_api_auth_login",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\flask_app.py",
        "x-source-line": 28,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\bauth\\b",
          "pii: \\bemail\\b",
          "Shadow API"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Express endpoint",
        "description": "Discovered from [express_app.js]:29\n\nFramework: Express\nLanguage: JavaScript\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_auth_register",
        "tags": [
          "Express",
          "JavaScript"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\express_app.js",
        "x-source-line": 29,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\btoken\\b",
          "auth: \\bauth\\b",
          "pii: \\bssn\\b",
          "pii: \\bemail\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [flask_app.py]:35\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_api_auth_register",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\flask_app.py",
        "x-source-line": 35,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\btoken\\b",
          "auth: \\bauth\\b",
          "pii: \\bssn\\b",
          "Shadow API"
        ]
      }
    },
...
    "/api/billing/invoices": {
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:60\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_api_billing_invoices",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 60,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\btoken\\b",
          "financial: \\bpayment\\b",
          "financial: \\bbilling\\b",
          "Shadow API"
        ]
      }
    },
    "/api/cart/add": {
      "post": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:77\n\nFramework: Flask\nLanguage: Python\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "post_api_cart_add",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 77,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/checkout/complete": {
      "post": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:81\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_checkout_complete",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 81,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "pii: \\bcredit.?card\\b",
          "financial: \\bpayment\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/payments/charge": {
      "post": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:55\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_payments_charge",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 55,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\btoken\\b",
          "pii: \\bemail\\b",
          "financial: \\bpayment\\b",
          "financial: \\bcharge\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/public/products": {
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:27\n\nFramework: Flask\nLanguage: Python\nRisk Level: INFO\nAuth Status: PUBLIC",
        "operationId": "get_api_public_products",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 27,
        "x-risk-level": "INFO",
        "x-auth-status": "PUBLIC",
        "security": []
      }
    },
...
    "/api/users/<int{user_id}>": {
      "put": {
        "summary": "Flask endpoint",
        "description": "Discovered from [flask_app.py]:47\n\nFramework: Flask\nLanguage: Python\nRisk Level: LOW\nAuth Status: PRIVATE",
        "operationId": "put_api_users_intuser_id",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\flask_app.py",
        "x-source-line": 47,
        "x-risk-level": "LOW",
        "x-auth-status": "PRIVATE",
        "x-risk-reasons": [
          "Mutation: PUT"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: user_id"
          }
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Flask endpoint",
        "description": "Discovered from [flask_app.py]:52\n\nFramework: Flask\nLanguage: Python\nRisk Level: LOW\nAuth Status: PRIVATE",
        "operationId": "delete_api_users_intuser_id",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\flask_app.py",
        "x-source-line": 52,
        "x-risk-level": "LOW",
        "x-auth-status": "PRIVATE",
        "x-risk-reasons": [
          "Mutation: DELETE"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: user_id"
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:44\n\nFramework: Flask\nLanguage: Python\nRisk Level: MEDIUM\nAuth Status: UNKNOWN",
        "operationId": "get_api_users_me",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 44,
        "x-risk-level": "MEDIUM",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\btoken\\b",
          "Shadow API"
        ]
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:26\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: HIGH\nAuth Status: UNKNOWN",
        "operationId": "get_api_users_id",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 26,
        "x-risk-level": "HIGH",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "pii: \\bemail\\b",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ]
      },
      "delete": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:56\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: HIGH\nAuth Status: PRIVATE",
        "operationId": "delete_api_users_id",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 56,
        "x-risk-level": "HIGH",
        "x-auth-status": "PRIVATE",
        "x-risk-reasons": [
          "admin: \\badmin\\b",
          "Mutation: DELETE"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ]
      },
      "put": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:48\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: LOW\nAuth Status: PRIVATE",
        "operationId": "put_api_users_id",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 48,
        "x-risk-level": "LOW",
        "x-auth-status": "PRIVATE",
        "x-risk-reasons": [
          "Mutation: PUT"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:49\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_api_users_user_id",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 49,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\btoken\\b",
          "pii: \\bemail\\b",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: user_id"
          }
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:73\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: HIGH\nAuth Status: PUBLIC",
        "operationId": "post_api_v1_auth_login",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 73,
        "x-risk-level": "HIGH",
        "x-auth-status": "PUBLIC",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "Sensitive PUBLIC"
        ],
        "security": [],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:81\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: HIGH\nAuth Status: PUBLIC",
        "operationId": "post_api_v1_auth_register",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 81,
        "x-risk-level": "HIGH",
        "x-auth-status": "PUBLIC",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "Sensitive PUBLIC"
        ],
        "security": [],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/charge": {
      "post": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:115\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: HIGH\nAuth Status: PRIVATE",
        "operationId": "post_api_v1_payments_charge",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 115,
        "x-risk-level": "HIGH",
        "x-auth-status": "PRIVATE",
        "x-risk-reasons": [
          "financial: \\bpayment\\b",
          "financial: \\bcharge\\b"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/payments/charge": {
      "post": {
        "summary": "MinimalAPI endpoint",
        "description": "Discovered from [Startup.cs]:142\n\nFramework: MinimalAPI\nLanguage: C#/.NET\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_v2_payments_charge",
        "tags": [
          "MinimalAPI",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Startup.cs",
        "x-source-line": 142,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "financial: \\bpayment\\b",
          "financial: \\bcharge\\b",
          "admin: \\badmin\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "summary": "Echo endpoint",
        "description": "Discovered from [go_app.go]:157\n\nFramework: Echo\nLanguage: Go\nRisk Level: MEDIUM\nAuth Status: UNKNOWN",
        "operationId": "post_auth_login",
        "tags": [
          "Echo",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 157,
        "x-risk-level": "MEDIUM",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Echo endpoint",
        "description": "Discovered from [go_app.go]:158\n\nFramework: Echo\nLanguage: Go\nRisk Level: MEDIUM\nAuth Status: UNKNOWN",
        "operationId": "post_auth_register",
        "tags": [
          "Echo",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 158,
        "x-risk-level": "MEDIUM",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}